finalize assignment 2
There were some errors and accidentally deleted files in the last push which was tagged as "solving errors of assignment 2". Tried to revert back to the correct version "updated assignment 2" but failed to do so. Hence, this commit is identical to "updated assignment 2" just to clear things up. Sorry for the confusion.